requirements.txt
commonroad-all
omegaconf
matplotlib<3.9
scipy==1.15.2
tqdm==4.66.4
ipykernel
